Decompose ReportScreen: extract isolated subscriptions
Replace COLLECTION.POLICY collection subscription with per-key lookup,
extract DeleteTransactionNavigateBackHandler and AccountManagerBanner
into isolated components to reduce re-renders (21 → 17 subscriptions).
