+++ src/pages/inbox/AccountManagerBanner.tsx
@@ -1,0 +1,67 @@
+import {useState} from 'react';
+import type {OnyxEntry} from 'react-native-onyx';
+import Banner from '@components/Banner';
+import {useMemoizedLazyExpensifyIcons} from '@hooks/useLazyAsset';
+import useLocalize from '@hooks/useLocalize';
+import useOnyx from '@hooks/useOnyx';
+import useThemeStyles from '@hooks/useThemeStyles';
+import getNonEmptyStringOnyxID from '@libs/getNonEmptyStringOnyxID';
+import Navigation from '@libs/Navigation/Navigation';
+import {getPersonalDetailsForAccountIDs} from '@libs/OptionsListUtils';
+import {getDisplayNameOrDefault} from '@libs/PersonalDetailsUtils';
+import {getParticipantsAccountIDsForDisplay, isConciergeChatReport} from '@libs/ReportUtils';
+import ONYXKEYS from '@src/ONYXKEYS';
+import ROUTES from '@src/ROUTES';
+import type * as OnyxTypes from '@src/types/onyx';
+
+type AccountManagerBannerProps = {
+    report: OnyxEntry<OnyxTypes.Report>;
+};
+
+function AccountManagerBanner({report}: AccountManagerBannerProps) {
+    const styles = useThemeStyles();
+    const {translate} = useLocalize();
+    const expensifyIcons = useMemoizedLazyExpensifyIcons(['Lightbulb']);
+    const [accountManagerReportID] = useOnyx(ONYXKEYS.ACCOUNT_MANAGER_REPORT_ID);
+    const [accountManagerReport] = useOnyx(`${ONYXKEYS.COLLECTION.REPORT}${getNonEmptyStringOnyxID(accountManagerReportID)}`);
+    const [personalDetails] = useOnyx(ONYXKEYS.PERSONAL_DETAILS_LIST);
+    const [isBannerVisible, setIsBannerVisible] = useState(true);
+
+    let chatWithAccountManagerText = '';
+    if (accountManagerReportID) {
+        const participants = getParticipantsAccountIDsForDisplay(accountManagerReport, false, true);
+        const participantPersonalDetails = getPersonalDetailsForAccountIDs([participants?.at(0) ?? -1], personalDetails);
+        const participantPersonalDetail = Object.values(participantPersonalDetails).at(0);
+        const displayName = getDisplayNameOrDefault(participantPersonalDetail);
+        const login = participantPersonalDetail?.login;
+        if (displayName && login) {
+            chatWithAccountManagerText = translate('common.chatWithAccountManager', `${displayName} (${login})`);
+        }
+    }
+
+    const dismissBanner = () => setIsBannerVisible(false);
+    const chatWithAccountManager = () => {
+        Navigation.navigate(ROUTES.REPORT_WITH_ID.getRoute(accountManagerReportID));
+    };
+
+    if (!accountManagerReportID || !isConciergeChatReport(report) || !isBannerVisible) {
+        return null;
+    }
+
+    return (
+        <Banner
+            containerStyles={[styles.mh4, styles.mt4, styles.p4, styles.br2]}
+            text={chatWithAccountManagerText}
+            onClose={dismissBanner}
+            onButtonPress={chatWithAccountManager}
+            shouldShowCloseButton
+            icon={expensifyIcons.Lightbulb}
+            shouldShowIcon
+            shouldShowButton
+        />
+    );
+}
+
+AccountManagerBanner.displayName = 'AccountManagerBanner';
+
+export default AccountManagerBanner;
